renamed parsing package to protocol
