+++ protocol/decode.go
@@ -1,4 +1,4 @@
-package parsing
+package protocol
 
 import (
 	"encoding/binary"
