+++ protocol/parsing_test.go
@@ -1,4 +1,4 @@
-package parsing
+package protocol
 
 import (
 	"bytes"
